overhaul architecture (WIP)
