+++ autopilot/contracts.go
@@ -1,8 +1,8 @@
 package autopilot
 
 import (
-	"go.sia.tech/renterd/api"
 	"go.sia.tech/renterd/internal/consensus"
+	"go.sia.tech/renterd/worker"
 	"go.sia.tech/siad/types"
 	"golang.org/x/crypto/blake2b"
 )
@@ -20,18 +20,22 @@ func (ap *Autopilot) deriveRenterKey(hostKey consensus.PublicKey) consensus.Priv
 	return pk
 }
 
-func (ap *Autopilot) defaultContracts() ([]api.Contract, error) {
-	cs, err := ap.renter.HostSetContracts("autopilot")
+func (ap *Autopilot) defaultContracts() ([]worker.Contract, error) {
+	cs, err := ap.bus.HostSetContracts("autopilot")
 	if err != nil {
 		return nil, err
 	}
-	contracts := make([]api.Contract, 0, len(cs))
+	contracts := make([]worker.Contract, 0, len(cs))
 	for _, c := range cs {
 		if c.ID == (types.FileContractID{}) || c.HostIP == "" {
 			continue
 		}
-		c.RenterKey = ap.deriveRenterKey(c.HostKey)
-		contracts = append(contracts, c)
+		contracts = append(contracts, worker.Contract{
+			HostKey:   c.HostKey,
+			HostIP:    c.HostIP,
+			ID:        c.ID,
+			RenterKey: ap.deriveRenterKey(c.HostKey),
+		})
 	}
 	return contracts, nil
 }
