+++ bus/client.go
@@ -1,13 +1,7 @@
-package api
+package bus
 
 import (
-	"bytes"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
-	"io/ioutil"
-	"net/http"
 	"time"
 
 	"go.sia.tech/jape"
@@ -15,8 +9,6 @@ import (
 	"go.sia.tech/renterd/internal/consensus"
 	"go.sia.tech/renterd/object"
 	rhpv2 "go.sia.tech/renterd/rhp/v2"
-	rhpv3 "go.sia.tech/renterd/rhp/v3"
-	"go.sia.tech/renterd/slab"
 	"go.sia.tech/renterd/wallet"
 	"go.sia.tech/siad/types"
 )
@@ -173,124 +165,6 @@ func (c *Client) RecordHostInteraction(hostKey PublicKey, i hostdb.Interaction) 
 	return
 }
 
-// RHPScan scans a host, returning its current settings.
-func (c *Client) RHPScan(hostKey PublicKey, hostIP string) (resp rhpv2.HostSettings, err error) {
-	err = c.c.POST("/rhp/scan", RHPScanRequest{hostKey, hostIP}, &resp)
-	return
-}
-
-// RHPPrepareForm prepares a contract formation transaction.
-func (c *Client) RHPPrepareForm(renterKey PrivateKey, hostKey PublicKey, renterFunds types.Currency, renterAddress types.UnlockHash, hostCollateral types.Currency, endHeight uint64, hostSettings rhpv2.HostSettings) (types.FileContract, types.Currency, error) {
-	req := RHPPrepareFormRequest{
-		RenterKey:      renterKey,
-		HostKey:        hostKey,
-		RenterFunds:    renterFunds,
-		RenterAddress:  renterAddress,
-		HostCollateral: hostCollateral,
-		EndHeight:      endHeight,
-		HostSettings:   hostSettings,
-	}
-	var resp RHPPrepareFormResponse
-	err := c.c.POST("/rhp/prepare/form", req, &resp)
-	return resp.Contract, resp.Cost, err
-}
-
-// RHPPrepareRenew prepares a contract renewal transaction.
-func (c *Client) RHPPrepareRenew(contract types.FileContractRevision, renterKey PrivateKey, hostKey PublicKey, renterFunds types.Currency, renterAddress types.UnlockHash, hostCollateral types.Currency, endHeight uint64, hostSettings rhpv2.HostSettings) (types.FileContract, types.Currency, types.Currency, error) {
-	req := RHPPrepareRenewRequest{
-		Contract:       contract,
-		RenterKey:      renterKey,
-		HostKey:        hostKey,
-		RenterFunds:    renterFunds,
-		RenterAddress:  renterAddress,
-		HostCollateral: hostCollateral,
-		EndHeight:      endHeight,
-		HostSettings:   hostSettings,
-	}
-	var resp RHPPrepareRenewResponse
-	err := c.c.POST("/rhp/prepare/renew", req, &resp)
-	return resp.Contract, resp.Cost, resp.FinalPayment, err
-}
-
-// RHPPreparePayment prepares an ephemeral account payment.
-func (c *Client) RHPPreparePayment(account rhpv3.Account, amount types.Currency, key PrivateKey) (resp rhpv3.PayByEphemeralAccountRequest, err error) {
-	req := RHPPreparePaymentRequest{
-		Account:    account,
-		Amount:     amount,
-		Expiry:     0, // TODO
-		AccountKey: key,
-	}
-	err = c.c.POST("/rhp/prepare/payment", req, &resp)
-	return
-}
-
-// RHPForm forms a contract with a host.
-func (c *Client) RHPForm(renterKey PrivateKey, hostKey PublicKey, hostIP string, transactionSet []types.Transaction) (rhpv2.Contract, []types.Transaction, error) {
-	req := RHPFormRequest{
-		RenterKey:      renterKey,
-		HostKey:        hostKey,
-		HostIP:         hostIP,
-		TransactionSet: transactionSet,
-	}
-	var resp RHPFormResponse
-	err := c.c.POST("/rhp/form", req, &resp)
-	return resp.Contract, resp.TransactionSet, err
-}
-
-// RHPRenew renews an existing contract with a host.
-func (c *Client) RHPRenew(renterKey PrivateKey, hostKey PublicKey, hostIP string, contractID types.FileContractID, transactionSet []types.Transaction, finalPayment types.Currency) (rhpv2.Contract, []types.Transaction, error) {
-	req := RHPRenewRequest{
-		RenterKey:      renterKey,
-		HostKey:        hostKey,
-		HostIP:         hostIP,
-		ContractID:     contractID,
-		TransactionSet: transactionSet,
-		FinalPayment:   finalPayment,
-	}
-	var resp RHPRenewResponse
-	err := c.c.POST("/rhp/renew", req, &resp)
-	return resp.Contract, resp.TransactionSet, err
-}
-
-// RHPFund funds an ephemeral account using the supplied contract.
-func (c *Client) RHPFund(contract types.FileContractRevision, renterKey PrivateKey, hostKey PublicKey, hostIP string, account rhpv3.Account, amount types.Currency) (err error) {
-	req := RHPFundRequest{
-		Contract:  contract,
-		RenterKey: renterKey,
-		HostKey:   hostKey,
-		HostIP:    hostIP,
-		Account:   account,
-		Amount:    amount,
-	}
-	err = c.c.POST("/rhp/fund", req, nil)
-	return
-}
-
-// RHPReadRegistry reads a registry value.
-func (c *Client) RHPReadRegistry(hostKey PublicKey, hostIP string, key rhpv3.RegistryKey, payment rhpv3.PayByEphemeralAccountRequest) (resp rhpv3.RegistryValue, err error) {
-	req := RHPRegistryReadRequest{
-		HostKey:     hostKey,
-		HostIP:      hostIP,
-		RegistryKey: key,
-		Payment:     payment,
-	}
-	err = c.c.POST("/rhp/registry/read", req, &resp)
-	return
-}
-
-// RHPUpdateRegistry updates a registry value.
-func (c *Client) RHPUpdateRegistry(hostKey PublicKey, hostIP string, key rhpv3.RegistryKey, value rhpv3.RegistryValue, payment rhpv3.PayByEphemeralAccountRequest) (err error) {
-	req := RHPRegistryUpdateRequest{
-		HostKey:       hostKey,
-		HostIP:        hostIP,
-		RegistryKey:   key,
-		RegistryValue: value,
-		Payment:       payment,
-	}
-	err = c.c.POST("/rhp/registry/update", req, nil)
-	return
-}
-
 // Contracts returns the current set of contracts.
 func (c *Client) Contracts() (contracts []rhpv2.Contract, err error) {
 	err = c.c.GET("/contracts", &contracts)
@@ -335,97 +209,9 @@ func (c *Client) SetHostSet(name string, hosts []consensus.PublicKey) (err error
 
 // HostSetContracts returns the latest contract for each host in the given set.
 // The ID and HostIP fields may be empty, depending on whether a contract exists
-// and a host announcement is known. The RenterKey field is always empty.
+// and a host announcement is known.
 func (c *Client) HostSetContracts(name string) (contracts []Contract, err error) {
 	err = c.c.GET(fmt.Sprintf("/hostsets/%s/contracts", name), &contracts)
-	return
-}
-
-// HostSetResolves returns the last announced IP for each host in the given set,
-// or the empty string if no announcement is known.
-func (c *Client) HostSetResolves(name string) (ips []string, err error) {
-	err = c.c.GET(fmt.Sprintf("/hostsets/%s/resolve", name), &ips)
-	return
-}
-
-// UploadSlab uploads data to a set of hosts. At most m*SectorSize bytes will be
-// read from src.
-func (c *Client) UploadSlab(src io.Reader, m, n uint8, height uint64, contracts []Contract) (s slab.Slab, err error) {
-	c.c.Custom("POST", "/slabs/upload", []byte{}, &s)
-
-	js, _ := json.Marshal(SlabsUploadRequest{
-		MinShards:     m,
-		TotalShards:   n,
-		Contracts:     contracts,
-		CurrentHeight: height,
-	})
-	body := io.MultiReader(bytes.NewReader(js), io.LimitReader(src, int64(m)*rhpv2.SectorSize))
-	req, err := http.NewRequest("POST", fmt.Sprintf("%v%v", c.c.BaseURL, "/slabs/upload"), body)
-	if err != nil {
-		panic(err)
-	}
-	req.SetBasicAuth("", c.c.Password)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return slab.Slab{}, err
-	}
-	defer io.Copy(ioutil.Discard, resp.Body)
-	defer resp.Body.Close()
-	if resp.StatusCode != 200 {
-		err, _ := ioutil.ReadAll(resp.Body)
-		return slab.Slab{}, errors.New(string(err))
-	}
-	err = json.NewDecoder(resp.Body).Decode(&s)
-	return
-}
-
-// DownloadSlab downloads data from a set of hosts.
-func (c *Client) DownloadSlab(dst io.Writer, s slab.Slice, contracts []Contract) (err error) {
-	c.c.Custom("POST", "/slabs/download", SlabsDownloadRequest{}, (*[]byte)(nil))
-
-	js, _ := json.Marshal(SlabsDownloadRequest{
-		Slab:      s,
-		Contracts: contracts,
-	})
-	req, err := http.NewRequest("POST", fmt.Sprintf("%v%v", c.c.BaseURL, "/slabs/download"), bytes.NewReader(js))
-	if err != nil {
-		panic(err)
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.SetBasicAuth("", c.c.Password)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return err
-	}
-	defer io.Copy(ioutil.Discard, resp.Body)
-	defer resp.Body.Close()
-	if resp.StatusCode != 200 {
-		err, _ := ioutil.ReadAll(resp.Body)
-		return errors.New(string(err))
-	}
-	_, err = io.Copy(dst, resp.Body)
-	return
-}
-
-// MigrateSlab migrates the specified slab.
-func (c *Client) MigrateSlab(s *slab.Slab, from, to []Contract, currentHeight uint64) (err error) {
-	req := SlabsMigrateRequest{
-		Slab:          *s,
-		From:          from,
-		To:            to,
-		CurrentHeight: currentHeight,
-	}
-	err = c.c.POST("/slabs/migrate", req, s)
-	return
-}
-
-// DeleteSlabs deletes the specified slabs.
-func (c *Client) DeleteSlabs(slabs []slab.Slab, contracts []Contract) (err error) {
-	req := SlabsDeleteRequest{
-		Slabs:     slabs,
-		Contracts: contracts,
-	}
-	err = c.c.POST("/slabs/delete", req, nil)
 	return
 }
 
@@ -461,8 +247,8 @@ func (c *Client) DeleteObject(name string) (err error) {
 	return
 }
 
-// NewClient returns a client that communicates with a renterd server listening
-// on the specified address.
+// NewClient returns a client that communicates with a renterd store server
+// listening on the specified address.
 func NewClient(addr, password string) *Client {
 	return &Client{jape.Client{
 		BaseURL:  addr,
