+++ worker/slab.go
@@ -1,10 +1,9 @@
-package main
+package worker
 
 import (
 	"context"
 	"io"
 
-	"go.sia.tech/renterd/api"
 	"go.sia.tech/renterd/slab"
 )
 
@@ -12,7 +11,7 @@ type slabMover struct {
 	pool *slab.SessionPool
 }
 
-func (sm slabMover) withHosts(ctx context.Context, contracts []api.Contract, fn func([]slab.Host) error) (err error) {
+func (sm slabMover) withHosts(ctx context.Context, contracts []Contract, fn func([]slab.Host) error) (err error) {
 	var hosts []slab.Host
 	for _, c := range contracts {
 		hosts = append(hosts, sm.pool.Session(c.HostKey, c.HostIP, c.ID, c.RenterKey))
@@ -39,7 +38,7 @@ func (sm slabMover) withHosts(ctx context.Context, contracts []api.Contract, fn 
 	return fn(hosts)
 }
 
-func (sm slabMover) UploadSlab(ctx context.Context, r io.Reader, m, n uint8, currentHeight uint64, contracts []api.Contract) (s slab.Slab, err error) {
+func (sm slabMover) UploadSlab(ctx context.Context, r io.Reader, m, n uint8, currentHeight uint64, contracts []Contract) (s slab.Slab, err error) {
 	sm.pool.SetCurrentHeight(currentHeight)
 	err = sm.withHosts(ctx, contracts, func(hosts []slab.Host) error {
 		s, err = slab.UploadSlab(r, m, n, hosts)
@@ -48,19 +47,19 @@ func (sm slabMover) UploadSlab(ctx context.Context, r io.Reader, m, n uint8, cur
 	return
 }
 
-func (sm slabMover) DownloadSlab(ctx context.Context, w io.Writer, s slab.Slice, contracts []api.Contract) error {
+func (sm slabMover) DownloadSlab(ctx context.Context, w io.Writer, s slab.Slice, contracts []Contract) error {
 	return sm.withHosts(ctx, contracts, func(hosts []slab.Host) error {
 		return slab.DownloadSlab(w, s, hosts)
 	})
 }
 
-func (sm slabMover) DeleteSlabs(ctx context.Context, slabs []slab.Slab, contracts []api.Contract) error {
+func (sm slabMover) DeleteSlabs(ctx context.Context, slabs []slab.Slab, contracts []Contract) error {
 	return sm.withHosts(ctx, contracts, func(hosts []slab.Host) error {
 		return slab.DeleteSlabs(slabs, hosts)
 	})
 }
 
-func (sm slabMover) MigrateSlab(ctx context.Context, s *slab.Slab, currentHeight uint64, from, to []api.Contract) (err error) {
+func (sm slabMover) MigrateSlab(ctx context.Context, s *slab.Slab, currentHeight uint64, from, to []Contract) (err error) {
 	sm.pool.SetCurrentHeight(currentHeight)
 	var fromHosts []slab.Host
 	for _, c := range from {
